+++ src/lib/components/SEO.svelte
@@ -30,11 +30,6 @@
 <svelte:head>
 	<title>{pageTitle}</title>
 	<meta name="description" content={metadescription} />
-	<meta
-		name="robots"
-		content="index, follow, max-snippet:-1, max-image-preview:large, max-video-preview:-1"
-	/>
-	<html lang={siteLanguage} />
 </svelte:head>
 
 <SvelteSeo